Add Favourites support
This adds a Favourite interface, implemented by Individual and Persona, and
backed by a D-Bus connection to the telepathy-logger service in
TpfPersonaStore.

The D-Bus logger proxy is a custom-written one in Vala, as the logger proxy
doesn't exist in telepathy-glib at the moment.

Whenever the is-favourite property of an Individual is changed, its new value
is propagated to _all_ of the individual's Personas. Conversely, if a single
Persona has is-favourite == true, the Individual containing that Persona will
also have is-favourite == true.