Add a key-file backend
This backend stores Personas in a key file which is completely under the
user's control, making it the only fully trusted Backend. Currently, it only
stores IM addresses, which can then be used as linkable properties by the
IndividualAggregator to link together Personas from multiple backends to
automagically produce Individuals.