Allow late linking of Individuals if Personas are added at runtime
If a Persona is added after the initial aggregation phase and it should be
linked to more than one existing Individual, the Individuals are now destroyed
and their Personas added to a new linked Individual containing the late
Persona.

This is necessary for supporting runtime linking and unlinking of Personas.