Block notification of Persona.is-favourite when changing it on the Individual
This fixes changing the favourite status of an Individual containing more than
one Persona, since previously the change wouldn't be applied. This is because
the change would be bubbled down to the first Persona, which would then
notify of the change to its is-individual property, causing the Individual's
is-favourite property to be re-calculated incorrectly, since the favourite
status of the _other_ Personas hadn't yet been changed. Closes: bgo#628108