Ensure existing Kf.Personas aren't overwritten when choosing new group names
If a key file contained several groups which weren't contiguously numbered on
startup, IDs for new groups would start from the first unused ID, and continue
happily into the next block of IDs which were already in use. This patch
ensures that if an ID is already in use, it won't be again for a new
Kf.Persona. Closes: bgo#628294