Fix saving the key file when it contains multi-byte characters
The calculation of the length of the data to be saved was being done in
characters, not bytes, whereas the I/O function was expecting a length in
bytes. This caused some number of bytes to be dropped from the end of the
file, potentially splitting a multibyte character and breaking the file
encoding (not to mention losing data otherwise). Closes: bgo#628930