Ensure a given individual isn't marked as added and replaced at the same time
When linking personas with Persona.is_user == true, it's possible for the
final_individuals during the call to add_personas() to be replaced and added
at the same time (i.e. added, then replaced). This breaks things, but can
be fixed by caching the set of added individuals before it's finalised, and
removing replaced individuals from it as we go.