Bug 645680 — Use random integer IDs for Kf.Personas
Using consecutive integer IDs meant that IDs often got re-used as Personas
were deleted and added, meaning that the same ID could, over time, be applied
to several different Personas which contained completely different IM
addresses. Any external references to such IDs which weren't updated over
time with the IDs themselves (such as the to-be-added anti-links) would end
up pointing to effectively the wrong Persona.

A perfect solution would be to hash the IM addresses in each Persona and use
that as the Persona's ID, but that's not possible as the set of IM addresses
in a Kf.Persona can change over time, while the ID has to remain constant.
Consequently, the use of a random integer evenly distributed over the space of
unsigned 32-bit integers will do.

Closes: bgo#645680