Convert the internal store of Personas in Individual to use HashSet
This converts the private variables in Individual which store its set of
Personas to use a single HashSet instead of a separate GLib.List and HashSet.

This won't compile, but making all the necessary adjustments to
Individual.personas to make it compile would've made the commit too complex
to review. The next commit will update Individual.personas.

Helps: bgo#640092