Allow generated source files to exist after distcleancheck.
It turns out to be nearly impossible to satisfy a standard distcleancheck
with Vala and its automake support. This is specifically a work-around
because there doesn't seem to be any obvious way to satisfy these
requirements:

  * the release tarball can be built without gobject-introspection, valac,
    or vapigen

  * it's safe to run "make distclean" in the release tarball without
    suddenly requiring the tools listed above (due to generated files
    being removed)

  * "make distcheck" doesn't fail its "distcleancheck" phase due to the
    following files remaining after distclean:

    - tests/lib/telepathy/contactlist/tp-test-contactlist.gir
    - tests/lib/telepathy/contactlist/tp-test-contactlist.vapi
    - backends/telepathy/lib/tp-lowlevel.gir
    - backends/telepathy/lib/tp-lowlevel.vapi

We've hit this last issue multiple times, and no fix seems to stick. So
stripping out this check seems to be the best option (until we can
remove the Vala/GObject-Introspection-free build requirement, when
they've stabilized more and we don't depend upon the latest versions which
may not be in all the major distros).