Bug 650422 — Add API for easily checking whether details are writeable
Add Persona.writeable_properties, listing the properties of that Persona
for which updated values will be stored. It's intended that the property
value will be constant for a given Persona subclass, but this isn't
guaranteed; it's possible that Persona subclasses may vary the value of
writeable-properties at run time in future.

Closes: bgo#650422