core: Add core anti-linking support
This adds the core of the anti-linking support, based around a new
AntiLinkable interface. This will be implemented by Persona subclasses which
can store anti-linking information (in the form of a set of Persona UIDs
which the given Persona should never be linked to).

This approach allows anti-linking information to be stored with the personas
(presumably in the primary persona store) and thus it should be network
transparent. i.e. Using folks on two different computers with a Google
Contacts address book as primary should cause the anti-linking data to be
shared.

This also includes the necessary IndividualAggregator changes.

Sadly, no unit tests are included.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=629537