aggregator: Ensure we don't remove Personas we're adding
If a two Personas with the same UID are in the added and removed sets for a
single call to personas_changed_cb() in the individual aggregator, make sure
that we remove the old instance and replace it with the new instance, instead
of just removing both instances. This covers the case where the real instance
of a (Tpf.) Persona disappears and is immediately replaced by its cached
copy.

NOTE: This needs more testing.