Bug 653777 — Add a helper function to create a writable persona
This adds IndividualAggregator.ensure_individual_property_writeable(), which
returns a persona which has the specified property writeable in the specified
individual. If no such persona exists already, a new one is created and
linked to the individual. If that all fails, an error is returned.

This allows for clients to write to properties of personas with confidence
that the updates will be written out to the backend stores. (Achieved by
calling IndividualAggregator.ensure_individual_property_writeable() first
and only writing to the property if that call succeeds.)

Closes: bgo#653777