core: Add PersonaStore:always-writeable-properties
This complements Persona:writeable-properties, listing the properties which
are guaranteed to always be writeable on the personas in a given persona
store. This is in contrast to Persona:writeable-properties, which may list
extra properties which are only writeable on that particular persona.

This could be the case with, for example, personas representing the user,
which might have extra writeable properties to allow the user to change their
alias or avatar.

Helps: bgo#653777