Use the key-file backend for writes in tests for non-writeable backends.
We could have used the EDS backend, but this helps keep the execution
of the tests as simple as possible.

This fixes a bug in the libsocialweb aggregation test where it was
using the EDS backend from the Aggregator default but its
FOLKS_BACKEND_STORE_KEY_FILE_PATH didn't include the EDS backend, so
it had nowhere to store its linking personas.