tracker: Fix the match-im-addresses test
Due to the Tracker tests now running with the tracker persona store as the
writeable store, Tracker personas in the tests are now linked by their
linkeable properties (as their persona store is now fully trusted due to
being the writeable store).

This caused the two personas in the match-im-addresses test to be linked
together implicitly by the aggregator, when the test case assumed they
weren't. This broke the test.

This commit fixes the test permanently by changing the way we consider
potential matches based on IM addresses. Previously, we only indicated a
potential match if the two individuals under consideration shared a common
IM address + protocol combination. This commit changes that so that two
individuals are now considered to potentially match if they share a common
IM address *only*. Protocols are no longer considered.

This allows us to change the way the personas are set up in the
match-im-addresses test so that they're no longer implicitly linked by the
aggregator.