libsocialweb: make PersonaStores immediately quiescent
Because we don't know if a libsocialweb service has 0
contacts or is not well configured or is not authenticated, etc.;
we might end up forever for contacts to arrive and never
achieve quiescence. Until libsocialweb provides us with more
information about the status of service (is it configured and
authenticated? How many contacts should we receive? etc.) we
immediately declare it quiescent after preparing it.