core: Be more thorough when removing individuals from the link map
Since the values of personas' linkable properties could've changed since we
added them to the link map, we have to be more thorough when removing
individuals from the map. We therefore now remove them by examining all the
values in the map, and removing every mapping to the individual we're
interested in.

Helps: bgo#656689