core: Fix individuals_changed_detailed mappings for unlinked individuals
It turns out that we can do a post-processing phase on the mappings to be
emitted by IndividualAggregator.individuals_changed_detailed and ensure that
individuals who've been unlinked are represented by a mapping from themselves
to the set of individuals replacing them, rather than a mapping from
themselves to null, and null to the set of individuals replacing them.

This probably isn't the most efficient way to fix the problem, but it works
and I can just about understand it.

Helps: bgo#656689, bgo#657282