Bug 656689 — Re-link personas on linkable properties being changed
Listen for notifications of changes to the linkable properties of every
persona in the IndividualAggregator. When we receive such a notification,
relink that persona by artificially removing them from the aggregator and
immediately re-adding them.

Like many things in the IndividualAggregator, this is ugly — but it appears
to work, and I can get my head around it.

Closes: bgo#656689