e-d-s: Make sure consecutive property changes work
Don't send any notification due to property changes until we've
done updating an entire persona, otherwise code that listens to
such notifies will see the persona in an intermediate state, whic
can cause problems. For instance, if such a callback tries to
change a property it might not recieve change notification for it.