Bug 658631 — crash at empathy startup
In the case that a persona is removed and the individual containing that
persona is replaced by a different non-null individual which doesn't contain
that persona (because it's been removed), we need to set the persona's
individual pointer to null rather than the replacement individual.

This stops us tripping the assertion in the setter for Persona.individual.

Closes: bgo#658631