individual: Prefer data from personas with writeable properties
As an addition to the fix for bgo#657783, if two personas come from stores
which have equal is-primary-store values, we now prefer the persona which
lists the given property as writeable. If both personas list (or don't list)
the property as writeable, we fall back to the given ordering function as
before.

Closes: bgo#657783