Deal with lingering callbacks in the eds/set-avatar test.
These can be called after we've set this._main_loop = null,
so if we don't check it first, we'll hit warnings (which get
treated as errors).

This isn't an ideal solution, but there doesn't seem to be a
great way to avoid the handlers from firing after we no
longer care (we need something like thaw/freeze_notify(),
except for general signals, not just notify).