eds: Reach quiescence in Edsf.PersonaStore for address books with no contacts
It’s possible for an address book to have the “do-initial-query” capability
yet not emit any ::objects-added signals on its view, if the address book has
no contacts.

This commit fixes that by listening for the ::complete signal on the address
book view and using that to determine quiescence instead.

This should theoretically also catch cases where the address book encounters
an error part-way through initialisation, since that will be reported via
the ::complete signal.