tests: Fix the EDS link-personas-diff-stores test
Due to the tightening up of checks for PersonaStores being prepared before
PersonaStore.add_persona_from_details() is called, this test was failing
because it didn’t wait for its persona stores to get prepared.

This fixes the problem by waiting for quiescence before launching into the
test code proper.