eds: Handle BUSY errors when opening address books better
If another process is in the middle of opening an address book when we come
to try and open it, our open() call will return immediately with a BUSY
error.

In this case, we need to wait for the EClient::opened signal to know when
the open operation has finished.

Hopefully, this will fix those irritating ‘blah is busy’ errors I frequently
get when running the EDS backend test suite. Hopefully it also won’t
introduce too many other problems.