eds: Cope with opening buggy EDS address book backends
Turns out that almost as soon as I’d pushed the previous commit, one of my
address books proved me wrong and returned successfully from the async open()
call without being open (according to is_opened()) afterwards. Bother.

This replaces the assertion that the above holds with some code which will
throw an error instead, which should be a bit more robust in the case of
buggy EDS address books.