inspect: Move readline() loop into the GLib main loop
The readline() loop is more pain than it’s worth, so we now use readline’s
callback interface for command line handling, allowing us to use a single
main loop and a single thread for both the aggregator and readline.

This should make signal handling a lot easier.

Note that this bumps our Vala requirement from 0.15.1 to 0.15.2.