Bug 660235 — Consistent signal handlers to a terminal would be cool
Fix folks-inspect’s handling of SIGINT, SIGTERM and EOF. In the former case,
we want to clear the current input buffer and display a new prompt. In
the case of SIGTERM we want to exit cleanly, and in the case of EOF we want
to exit if the input buffer is empty.

The combination of readline, Unix signals and threads in this was unfun.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=660235