Match names better for PotentialMatches
We match alias as well as full name, to e.g. match an EDS contact (full name)
with an IM contact (alias).

Also, we make exact name matches bring the priority to HIGH alone, which
is needed in the IM case since the alias is the only piece of data
we have to match on.

https://bugzilla.gnome.org/show_bug.cgi?id=668499