libsocialweb: Tidy up Swf.PersonaStore.prepare() to return asynchronously
Previously, the prepare() method would return synchronously, firing off
some asynchronous methods in libsocialweb, which would later return in
callbacks and (theoretically) cause the persona store to reach quiescence.

This behaviour was not playing nicely with the expected asynchronous
behaviour of prepare(), and wasn't handling errors properly — on an error,
the code would cancel preparation, but not throw an error or remove the
persona store.

The code now throws errors and removes the persona store if preparation
fails, which should hopefully help some of the not-reaching-quiescence issues
we’ve been seeing.

Helps: https://bugzilla.gnome.org/show_bug.cgi?id=670191