telepathy: Fix weak_unref() warnings
If a TpContact is destroyed and our weak reference handler is invoked, it
could previously not remove the TpContact's pointer from the map of weak
references held by the persona store. When the persona was later destroyed,
it would call weak_unref() on the stale pointer and cause a warning.

This patch ensures we always remove pointers from the map when their
weak notification functions are called.