core: Replace the linking step in IA.ensure_individual_property_writeable()
It’s a waste of time to create a new Persona to ensure the property is
writeable, then create _another_ new Persona to link it to the Individual.
Why not just create the writeable Persona so that it’s automatically linked
to the Individual?

Helps: https://bugzilla.gnome.org/show_bug.cgi?id=629537