Bug 677166 — Salut personas survive disconnection
Handle TpAccounts being disabled by listening for
TpAccountManager::account-disabled rather than (erroneously) assuming that
TpAccount:enabled will have been changed by the time the corresponding
TpConnection is disconnected.

This fixes a race condition when accounts are disabled, where the account’s
personas would persist if TpAccount:enabled hadn’t changed by the time the
connection was disconnected.

This comes at the cost of potentially storing and re-loading the set of
personas for that account to the cache, only to later delete the cache file
when TpAccount:enabled changes. I can’t think of a simple fix for this.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=677166