eds: Fix invalid cast in Edsf.PersonaStore
E.SourceRegistry.find_extension() always returns an E.Source — we must then
call get_extension() on that E.Source to get the relevant extension. This
commit fixes an invalid cast here and factors out code to check for backend
name matches for things like the ‘google’ and ‘ldap’ address book backends.

Note that this depends on a new GIR annotation in EDataServer.