core: Change default EDS PersonaStore ID to ‘system-address-book’
The new version of EDS uses that instead of ‘system’. This shouldn’t pose
upgrade problems (e.g. if a user has ‘system’ stored in GSettings), since in
that case the ‘system’ persona store wouldn’t be found, and the fallback
would be to use the persona store marked as the system-set default — which
is the ‘system-address-book’ persona store, as required.