eds: Add an Edsf.Persona.in_google_personal_group property
This is set whenever the persona is in the “My Contacts” group in Google
Contacts, which is normally exposed by EDS as being in the “Personal”
category. This system group can’t be reliably detected normally as “Personal”
is translated by EDS. This caused problems for gnome-contacts, hence the
new API.

The new API requires EDS 3.5.3 at runtime, and will always be false
otherwise.

Helps: https://bugzilla.gnome.org/show_bug.cgi?id=676383