Bug 681726 — empathy crashed with SIGABRT in g_assertion_message()
Handle the case where calling a change_prop() method on an Individual
doesn’t find a suitable Persona to change the property on. Previously, the
code assumed that not successfully changing the property meant an error
must’ve been set. This was obviously wrong.

This patch changes it to only try and throw an error if an error has actually
been set. If no suitable Personas exist to write a property change to, the
property change will fail with PropertyError.NOT_WRITEABLE.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=681726