telepathy: Handle connections disappearing during a ‘yield’ while connecting
It seems it’s possible for the TpConnection to disconnect while in an async
function call below the _notify_connection_cb(). Handle this possibility
gracefully, rather than exploding.

Note: This fix is basically a guess. The crash can’t be reproduced, and
no debug output was available from folks for the crash.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=683093