eds: handle removal of store without removal of personas
The assumption in the IndividualAggregator that all backends remove
their personas before removing a store did not hold for the EDS
backend when stores were removed via set_persona_stores() or
disable_store().

Fixing that in EDS is tricky, so better make the IndividualAggregator
more resilient and remove any remaining personas when the store gets
removed.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=689146