telepathy: Yield subsequent Logger.prepare() calls if one is pending
If one PersonaStore began to prepare its Logger, then yielded on a D-Bus
call, another PersonaStore could begin to prepare its Logger and also
try to create a static D-Bus proxy (which is only supposed to be created
once). This is because the original async call hasn’t returned and set
Logger._logger by that point.

Fix this by keeping a queue of pending prepare() calls which is signalled
by the initial prepare() call once it finishes yielding.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=677633