eds test: add test for specific behavior of linkable properties
This splits out a subtle and complicated linkable-properties bug discovered
by the set-emails test where changing the emails (which happen to be a
linkable property in the EDS backend) causes an unnecessary Edsf.Persona
creation and removal cycle.

The set-* tests are meant to confirm that reading and writing the given
property works, so seperating this should let us simplify those tests
slightly yet focus on the secondary details we care about in a specific test
case.