Bug 691382 — Restore tests disabled by bug #668415 (which has been fixed)
This ports the store-removed test to the current EDS introspection bindings
and removes the change-primary-store test on the basis that we don't really
support clients changing the primary store.

In the future, we may want to add a test to ensure we cope with the primary
store disappearing, but that's a different case.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=691382