support geo location information
This adds "Location" class with latitude and longitude and
single-valued "LocationDetails" which exposes that information for a
contact. The logic for choosing "the best" among many locations from
different personas is left out for now. Instead folks picks according
to the default ordering of single values, which prefers values from
the primary store.

The EDS backend provides, creates and writes the GEO property as
location.