core: Connect to Persona::notify, not all its details, in Individual
In practice, we care about basically every property anyway, and
tests/eds/perf was spending more than 6% of its time connecting
to these signals. Just connect to notify and demultiplex through a
lookup table for about a 5% speed-up.

In principle we could make the lookup table into a
hash table if it proves to be worth it, but life is probably too short.

https://bugzilla.gnome.org/show_bug.cgi?id=697209