telepathy: Fix build with --disable-zeitgeist
Vala doesn’t pass #if conditionals through to generated C code, so when
building folks from a tarball, the --[enable|disable]-zeitgeist option was
having no effect, and Zeitgeist was always being pulled into the build.

Fix this by separating the Zeitgeist code out into a small noinst shim
library and always generating C code for two versions of it: one with
Zeitgeist enabled, and one with it disabled. The correct one’s C code is
then compiled and linked into libfolks-telepathy.la when building from the
tarball.

This is a horrible, ugly fix, but the only better fix is to disable
distribution of generated C code. This was attempted, but thwarted by
automake’s Vala support.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=701099