build: distcheck fails trying to remove a srcdir file
When we referenced $(srcdir)/libfolks_telepathy_la_vala.stamp, we were
telling automake that it was OK to delete it from the srcdir, which is
considered bad form. By referring to it from the builddir, we don't
implicitly suggest it can be deleted (as it is as a part of its creation
process).

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=703830